requirements.txt
Django
djangorestframework
django-filter
graphene-django